move math specific testings into QuantaTools/maths_test_questions.py
