+++ QuantaTools/__init__.py
@@ -2,7 +2,6 @@
 from .model_config import ModelConfig
 from .model_token_to_char import token_to_char, tokens_to_string
 from .model_loss import logits_to_tokens_loss, loss_fn
-from .model_testing import test_correctness_on_num_questions
 
 # useful_*.py: Contains data on the useful token positions and useful nodes (attention heads and MLP neurons) that the model uses in predictions
 from .useful_config import UsefulConfig 
@@ -40,4 +39,6 @@ from .maths_constants import MathsToken, MathsBehavior, MathsAlgorithm
 from .maths_utilities import set_maths_vocabulary, set_maths_question_meanings, int_to_answer_str, tokens_to_unsigned_int, tokens_to_answer, insert_question_number, make_a_maths_question_and_answer
 from .maths_complexity import get_maths_question_complexity, get_maths_min_complexity, calc_maths_quanta_for_position_nodes
 from .maths_data_generator import maths_data_generator_core, maths_data_generator, make_maths_questions_and_answers
-from .maths_test_questions import make_maths_test_questions_and_answers, test_maths_questions_by_complexity, test_maths_questions_by_impact, test_maths_questions_and_add_useful_node_tags, TRICASE_QUESTIONS, make_maths_tricase_questions
+from .maths_test_questions import make_maths_test_questions_and_answers, test_maths_questions_by_complexity, \
+    test_maths_questions_by_impact, test_maths_questions_and_add_useful_node_tags, TRICASE_QUESTIONS, \
+    make_maths_tricase_questions, test_correctness_on_num_questions
