+++ QuantaTools/maths_test_questions.py
@@ -1,7 +1,9 @@
 import random
 import torch
 import transformer_lens.utils as utils
+from tqdm.notebook import tqdm
 
+from . import acfg, maths_data_generator
 from .model_loss import logits_to_tokens_loss, loss_fn
 from .model_token_to_char import tokens_to_string
 
@@ -619,3 +621,58 @@ def make_maths_tricase_questions(cfg):
             t_questions = make_maths_tricase_questions_core(cfg, answer_digit, operation)
             # Use a tuple of (answer_digit, operation) as the key for indexing
             cfg.tricase_questions_dict[(answer_digit, operation)] = t_questions
+
+
+def test_correctness_on_num_questions(cfg, num_questions=1000000):
+  store_perc_sub = cfg.perc_sub
+  store_perc_mult = cfg.perc_mult
+
+  def print_config():
+      print("%Mult=", cfg.perc_mult, "%Sub=", cfg.perc_sub, "%Add=", cfg.perc_add(), "File", cfg.file_config_prefix())
+
+  print_config()
+  print()
+
+  if cfg.perc_add() > 0:
+    print("Addition:")
+    cfg.perc_sub = 0
+    cfg.perc_mult = 0
+    test_correctness_on_num_questions_core(cfg, num_questions=num_questions)
+
+  if store_perc_sub > 0:
+    print("Subtraction:")
+    cfg.perc_sub = 100
+    cfg.perc_mult = 0
+    test_correctness_on_num_questions_core(cfg)
+    print()
+
+  cfg.perc_sub = store_perc_sub
+  cfg.perc_mult = store_perc_mult
+
+
+def test_correctness_on_num_questions_core(cfg, num_questions=1000000):
+  acfg.verbose = False
+
+  cfg.analysis_seed = 345621  # Randomly chosen
+  local_ds = maths_data_generator(cfg=cfg)  # Re-initialise the data generator
+
+  the_successes = 0
+  the_fails = 0
+
+  num_batches = num_questions//cfg.batch_size
+  for epoch in tqdm(range(num_batches)):
+      tokens = next(local_ds)
+
+      the_fails = test_maths_questions_by_impact(cfg, acfg, tokens, 0, False)
+
+      if the_fails>0:
+        break
+
+      the_successes = the_successes + cfg.batch_size
+
+      if epoch % 100 == 0:
+          print("Batch", epoch, "of", num_batches, "#Successes=", the_successes)
+
+  print("successes", the_successes, "num_fails", the_fails)
+  if the_fails > 0:
+    "WARNING: Model is not fully accurate. It failed the 1M Q test"
